240611 update
-Narration Script Update
