+++ Assets/Scripts/MR/NarrationControl_FamilyPicture.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a286bece314fc4549872ea5d27069a46
+guid: 5da6264e035ef65478f008315389db1b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
