+++ Assets/Scripts/MR/Test/MR_SetObjectActive.cs
@@ -2,13 +2,10 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class MR_SetActive : MonoBehaviour
+public class MR_SetObjectActive : MonoBehaviour
 {
-   
-
     void Start()
     {
         gameObject.SetActive(false);
-
     }
 }
